Add a simple landing page separate to the reading list.
