+++ middleware.ts
@@ -1,6 +1,8 @@
 import { authMiddleware } from '@clerk/nextjs'
 
-export default authMiddleware()
+export default authMiddleware({
+  publicRoutes: ['/'],
+})
 
 export const config = {
   matcher: ['/((?!.*\\..*|_next).*)', '/', '/(api|trpc)(.*)'],
