+++ pages/index.tsx
@@ -1,51 +1,20 @@
 import { Inter } from 'next/font/google'
 import Head from 'next/head'
-import { BookListBook, BookList } from '@/components/BookList'
-import { PrismaClient } from '@prisma/client'
-import { UserButton } from '@clerk/nextjs'
-import { buildClerkProps, getAuth } from '@clerk/nextjs/server'
-import { GetServerSideProps } from 'next'
+import Link from 'next/link'
 
 const inter = Inter({ subsets: ['latin'] })
 
-export default function Home({ books }: { books: BookListBook[] }) {
+export default function Home() {
   return (
-    <main className={`${inter.className}`}>
+    <main
+      className={`${inter.className} flex flex-col items-center justify-center overflow-hidden`}
+      style={{ height: '100dvh' }} /* Tailwind doesn't have dvh atm */
+    >
       <Head>
         <title>My books</title>
       </Head>
-      <div className="ml-auto w-fit p-3 pb-0">
-        <UserButton afterSignOutUrl="/" />
-      </div>
-      <h1 className="mx-auto mt-6 w-fit text-2xl">My books</h1>
-      <BookList books={books} />
+      <h1 className="text-2xl">My books</h1>
+      <Link href="/dashboard">Dashboard</Link>
     </main>
   )
-}
-
-// Needs to be SSRd because it can vary based on signed-in user.
-export const getServerSideProps: GetServerSideProps = async ({ req }) => {
-  const { userId } = getAuth(req)
-  if (userId == null) {
-    throw new Error('Should not have access to this page when not signed in')
-  }
-
-  const prisma = new PrismaClient()
-  const books: BookListBook[] = await prisma.book.findMany({
-    select: {
-      id: true,
-      title: true,
-      author: true,
-    },
-    where: {
-      userId,
-    },
-  })
-
-  return {
-    props: {
-      books,
-      ...buildClerkProps(req),
-    },
-  }
 }
